requirements.txt
google-api-python-client==2.96.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-telegram-bot==13.15
python-dotenv==1.0.0
httplib2==0.22.0
oauth2client==4.1.3
gspread==5.11.0
